Update Libc Functions (Mandatory)

+++ ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: harndt <humberto.arndt@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/04 19:18:23 by harndt            #+#    #+#             */
-/*   Updated: 2022/04/21 17:18:30 by harndt           ###   ########.fr       */
+/*   Updated: 2022/04/25 15:56:30 by harndt           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *str, int c, int n)
+void	*ft_memset(void *str, int c, size_t n)
 {
 	unsigned char	*ptr;
